Added tests for gender type

+++ pokemon_test.go
@@ -17,7 +17,7 @@ func TestStringNoNickname(t *testing.T) {
         Gender: Female,
         Level: 5,
         CurrentHP: 11,
-        Stats: [6]uint8{11, 6, 5, 6, 5, 5},
+        Stats: [6]uint{11, 6, 5, 6, 5, 5},
     }
 
     if (p.String() != "Bulbasaur\u2640\tLv5\nHP: 11/11\n") {
